% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphalt.R
\docType{data}
\name{alphalt}
\alias{alphalt}
\title{Line x Tester data for Alpha Lattice}
\format{
A data frame of five variables of 15 crosses derived from five lines and three testers.
\describe{
 \item{replication}{four replications}
 \item{block}{ten blocks}
 \item{line}{five inbred genotype}
 \item{tester}{three inbred genotype}
 \item{yield}{trait of intrest}

}
}
\usage{
data(alphalt)
}
\description{
The sample data containing only crosses laid out in Alpha Lattice design
}
\examples{
results1 = ltcalpha(alphalt, replication, block, line, tester, yield)
}
\keyword{datasets}
