% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaltc.R
\docType{data}
\name{alphaltc}
\alias{alphaltc}
\title{Line x Tester data (only Crosses) in Alpha Lattice design.}
\format{
A data frame of five variables of 15 crosses derived from five lines and three testers.
\describe{
 \item{replication}{four replications}
 \item{block}{ten blocks}
 \item{line}{five inbred genotype}
 \item{tester}{three inbred genotype}
 \item{yield}{trait of intrest}

}
}
\usage{
data(alphaltc)
}
\description{
The Line x Tester data of containing only crosses laid out in Alpha Lattice design.
}
\examples{
 result = ltc(alphaltc, replication, line, tester, yield, block)
}
\seealso{
\code{\link{rcbdltc}}
   ,\code{\link{alphaltcchk}}
   ,\code{\link{rcbdltcchk}}
}
\keyword{datasets}
