% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaltcchk.R
\docType{data}
\name{alphaltcchk}
\alias{alphaltcchk}
\title{Line x Tester data (Crosses and Checks) in Alpha Lattice}
\format{
A dataframe of six variables.
\describe{
 \item{replication}{three replications}
 \item{block}{six blocks}
 \item{line}{five lines}
 \item{tester}{three testers}
 \item{check}{three check}
 \item{yield}{trait of intrest}
}
}
\usage{
data(alphaltcchk)
}
\description{
The sample Line x Tester data of containing crosses and checks laid out in Alpha Lattice design. The data is composed of five lines, three testers
and three checks.
}
\examples{
result = ltcchk(alphaltcchk, replication, line, tester, check, yield, block)
}
\seealso{
\code{\link{rcbdltc}}
   ,\code{\link{alphaltc}}
   ,\code{\link{rcbdltcchk}}
}
\keyword{datasets}
