% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaltcs.R
\docType{data}
\name{alphaltcs}
\alias{alphaltcs}
\title{Line x Tester data (only Crosses) with single plant observations laid in Alpha Lattice design.}
\format{
A data frame of 15 crosses derived from five lines and three testers.
\describe{
 \item{replication}{four replications}
 \item{block}{five blocks}
 \item{line}{five inbred genotype}
 \item{tester}{three inbred genotype}
 \item{obs}{four single plant observations}
 \item{yield}{yield as a dependent trait}
}
}
\usage{
data(alphaltcs)
}
\description{
The Line x Tester data containing single plant observations of only crosses laid out in Alpha Lattice design.
}
\examples{
 result = ltcs(alphaltcs, replication, line, tester, obs, yield, block)
}
\seealso{
\code{\link{rcbdltcs}}
   ,\code{\link{alphaltcchk}}
   ,\code{\link{rcbdltcchk}}
   ,\code{\link{rcbdltcmt}}
}
\keyword{datasets}
