% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\docType{data}
\name{check}
\alias{check}
\title{Commercial check data}
\format{
A data frame of 7 variables of 3  maize Commercial checks.
\describe{
 \item{Check}{Name of the check}
 \item{DTP}{Days to anthesis}
 \item{DTS}{Days to silking}
 \item{DM}{Days to maturity}
 \item{PH}{Plant Height}
 \item{EH}{Ear Height}
 \item{GY}{DGrain yield}
}
}
\usage{
data(check)
}
\description{
The sample data containing mean values of 3 maize Commercial checks for 7 variables.
}
\examples{
\dontrun{# Standard Heterosis
library(gpbStat)
data(hybrid)
data(check)
df = hcc(hybrid, check)
df
}
}
\seealso{
\code{\link{alphaltcchk}}
   ,\code{\link{alphaltc}}
   ,\code{\link{rcbdltcchk}}
}
\keyword{datasets}
