% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm2alpha.R
\docType{data}
\name{dm2alpha}
\alias{dm2alpha}
\title{Diallel Method 2 data in Alpha Lattice.}
\format{
A data frame for Diallel analysis Method 2 containing four variables of 105 crosses and 15 parents.
\describe{
 \item{rep}{four replications}
 \item{blk}{twelve blocks}
 \item{parent1}{five inbred genotype}
 \item{parent2}{three inbred genotype}
 \item{TW}{data for test weight}

}
}
\usage{
data(dm2alpha)
}
\description{
The Diallel Method 2 data laid out in Alpha Lattice Design.
}
\examples{
 result2 =  dm2(dm2alpha, rep, parent1, parent2, TW, blk)
}
\seealso{
\code{\link{alphaltcchk}}
   ,\code{\link{alphaltc}}
   ,\code{\link{rcbdltcchk}}
   ,\code{\link{dm2rcbd}}
}
\keyword{datasets}
