% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm2rcbd.R
\docType{data}
\name{dm2rcbd}
\alias{dm2rcbd}
\title{Diallel Method 2 data in RCBD}
\format{
A data frame for Diallel analysis Method 2 containing four variables of 105 crosses and 15 parents.
\describe{
 \item{rep}{four replications}
 \item{parent1}{five inbred genotype}
 \item{parent2}{three inbred genotype}
 \item{DTP}{data for days to pollen shed}

}
}
\usage{
data(rcbdltc)
}
\description{
The Diallel Method 2 data laid out in Randomized Complete Block Design (RCBD).
}
\examples{
 result2 =  dm2(dm2rcbd, rep, parent1, parent2, DTP)
}
\seealso{
\code{\link{alphaltcchk}}
   ,\code{\link{alphaltc}}
   ,\code{\link{rcbdltcchk}}
   ,\code{\link{dm2alpha}}
}
\keyword{datasets}
