% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\docType{data}
\name{hybrid}
\alias{hybrid}
\title{Experimental Hybrid data}
\format{
A data frame of 7 variables of 65 experimental maize hybrids.
\describe{
 \item{Hybrid}{Name of the hybrid}
 \item{DTP}{Days to anthesis}
 \item{DTS}{Days to silking}
 \item{DM}{Days to maturity}
 \item{PH}{Plant Height}
 \item{EH}{Ear Height}
 \item{GY}{DGrain yield}
}
}
\usage{
data(hybrid)
}
\description{
The sample data containing mean values of 65 experimental maize hybrids for 7 variables.
}
\examples{
\dontrun{# Standard Heterosis
library(gpbStat)
data(hybrid)
data(check)
df = hcc(hybrid, check)
df
}
}
\seealso{
\code{\link{alphaltcchk}}
   ,\code{\link{alphaltc}}
   ,\code{\link{rcbdltcchk}}
}
\keyword{datasets}
