% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltcaplha.R
\name{ltcalpha}
\alias{ltcalpha}
\title{Analysis of Line x Tester data laid out in Alpha Lattice experimental design}
\usage{
ltcalpha(data, block, replication, line, tester, y)
}
\arguments{
\item{data}{dataframe containing following variables}

\item{block}{number of blocks per replication}

\item{replication}{number of replications}

\item{line}{number of lines}

\item{tester}{number of testers}

\item{y}{a trait of interest}
}
\value{
Results of line by tester analysis
}
\description{
Used for analyzing the line x tester data containing only crosses laid out in alpha lattice design
}
\details{
Analyzing the line by tester data only using the data from crosses which are evaluated in alpha lattice design.
}
\examples{
library(gpbStat)
data(alphalt)
results1 = ltcalpha(alphalt, replication, block, line, tester, yield)
results1
}
\author{
Nandan Patil
}
