% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltcrcbd.R
\name{ltcrcbd}
\alias{ltcrcbd}
\title{Analysis of Line x Tester data laid out in RCBD experimental design}
\usage{
ltcrcbd(data, replication, line, tester, y)
}
\arguments{
\item{data}{dataframe containing following variables}

\item{replication}{number of replications}

\item{line}{number of lines}

\item{tester}{number of testers}

\item{y}{trait of interest}
}
\value{
Results of line by tester analysis
}
\description{
Used for analyzing the line x tester data containing only crosses laid out in RCBD design
}
\details{
Analyzing the line by tester data containing only crosses which are evaluated in Randomized Complete Block Design (RCBD).
}
\examples{
library(gpbStat)
data(rcbdlt)
results2 = ltcrcbd(rcbdlt, replication, line, tester, yield)
results2
}
\author{
Nandan Patil
}
