% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcbdlt.R
\docType{data}
\name{rcbdlt}
\alias{rcbdlt}
\title{Line x Tester data for RCBD}
\format{
A data frame of four variables of 15 crosses derived from five lines and three testers.
\describe{
 \item{replication}{four replications}
 \item{line}{five inbred genotype}
 \item{tester}{three inbred genotype}
 \item{yield}{trait of intrest}

}
}
\usage{
data(rcbdlt)
}
\description{
The sample data containing only crosses laid out in RCBD.
}
\examples{
results2 = ltcrcbd(rcbdlt, replication, line, tester, yield)
}
\keyword{datasets}
