% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcbdltc.R
\docType{data}
\name{rcbdltc}
\alias{rcbdltc}
\title{Line x Tester data in RCBD}
\format{
A data frame of four variables of 15 crosses derived from five lines and three testers.
\describe{
 \item{replication}{four replications}
 \item{line}{five inbred genotype}
 \item{tester}{three inbred genotype}
 \item{yield}{trait of intrest}

}
}
\usage{
data(rcbdltc)
}
\description{
The sample Line x Tester data containing only crosses laid out in Randomized Complete Block Design (RCBD).
}
\examples{
 result = ltc(rcbdltc, replication, line, tester, yield)
}
\seealso{
\code{\link{alphaltcchk}}
   ,\code{\link{alphaltc}}
   ,\code{\link{rcbdltcchk}}
}
\keyword{datasets}
