% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcbdltcchk.R
\docType{data}
\name{rcbdltcchk}
\alias{rcbdltcchk}
\title{Line x Tester data (Crosses and Checks) in RCBD}
\format{
A dataframe of six variables.
\describe{
 \item{replication}{four replications}
 \item{line}{five lines}
 \item{tester}{three testers}
 \item{yield}{trait of intrest}

}
}
\usage{
data(rcbdltcchk)
}
\description{
The sample Line x Tester data of containing crosses and checks laid out in Randomized Complete Block Design (RCBD). The data is composed of five lines, three testers
and three checks.
}
\examples{
result = ltcchk(rcbdltcchk, replication, line, tester, check, yield)
}
\seealso{
\code{\link{rcbdltc}}
   ,\code{\link{alphaltc}}
   ,\code{\link{alphaltcchk}}
}
\keyword{datasets}
