% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcbdltcs.R
\docType{data}
\name{rcbdltcs}
\alias{rcbdltcs}
\title{Line x Tester data (only Crosses) with single plant observations laid in RCBD design.}
\format{
A data frame of 15 crosses derived from five lines and three testers.
\describe{
 \item{replication}{four replications}
 \item{line}{five inbred genotype}
 \item{tester}{three inbred genotype}
 \item{obs}{four single plant observations}
 \item{yield}{yield as a dependent trait}
}
}
\usage{
data(rcbdltcs)
}
\description{
The Line x Tester data containing single plant observations of only crosses laid out in RCBD design.
}
\examples{
 result = ltcs(rcbdltcs, replication, line, tester, obs, yield)
}
\seealso{
\code{\link{rcbdltcs}}
   ,\code{\link{alphaltcchk}}
   ,\code{\link{rcbdltcchk}}
   ,\code{\link{rcbdltcmt}}
}
\keyword{datasets}
