% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm2alpha.R
\docType{data}
\name{dm2alpha}
\alias{dm2alpha}
\title{Diallel Method 2 data in Alpha Lattice.}
\format{
A data frame for Diallel analysis Method 2 containing 105 crosses and 15 parents.
\describe{
 \item{replication}{two replications}
 \item{block}{twelve blocks}
 \item{parent1}{fifteen inbred genotype}
 \item{parent2}{fifteen inbred genotype}
 \item{TW}{data for test weight}

}
}
\usage{
data(dm2alpha)
}
\description{
The Diallel Method 2 data laid out in Alpha Lattice Design.
}
\examples{
 result2 =  dm2(dm2alpha, replication, parent1, parent2, TW, block)
}
\seealso{
\code{\link{alphaltcchk}}
   ,\code{\link{alphaltc}}
   ,\code{\link{rcbdltcchk}}
   ,\code{\link{dm2rcbd}}
}
\keyword{datasets}
