% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dti.R
\name{dti}
\alias{dti}
\title{Estimation of Drought Tolerance Indices.}
\usage{
dti(data, environment, genotype, traits, ns, st)
}
\arguments{
\item{data}{dataframe containing following variables}

\item{environment}{column with two levels i.e., non-stress and stress conditions}

\item{genotype}{genotypes evaluated}

\item{traits}{trait of interest}

\item{ns}{name of level indicating evaluation under non-stress (irrigated) conditions}

\item{st}{name of level indicating evaluation under stress conditions}
}
\value{
\item{\code{TOL}}{Stress tolerance.}\item{\code{STI}}{Stress tolerance index.}\item{\code{SSPI}}{Stress susceptibility percentage index.}\item{\code{YI}}{Yield index.}\item{\code{YSI}}{Yield stability index.}\item{\code{RSI}}{Relative
stress index.}\item{\code{MP}}{Mean productivity.}\item{\code{GMP}}{Geometric mean productivity}\item{\code{HM}}{Harmonic mean.}\item{\code{MRP}}{Mean relative performance.}\item{\code{PYR}}{Percent yield Reduction.}\item{\code{PYR}}{Drought
Susceptibility Index.}\item{\code{SSP}}{Stress Susceptibility Index.}
}
\description{
Estimation of Drought Tolerance Indices.
}
\details{
Estimation various Drought Tolerance Indices of genotypes evaluated under stress and non-stress conditions of both replicated and non-replicated data.
}
\note{
The function can handle both replicated and non-replicated data refer the examples.
}
\examples{
\dontrun{# Estimating drought tolerance indices
library(gpbStat)

data(datdti)
result1 =  dti(datdti, environment = ENV, genotype = GEN, datdti[,3:8],
              ns = 'NS-DWR', st = 'ST-DWR')
result1

data(datrdti)
result2 = dti(datrdti, environment = ENV, genotype = GEN, datrdti[,4:9],
            ns = 'NS-DWR', st = 'ST-DWR')
result2
}
}
\references{
Pour‐Aboughadareh, A., Yousefian, M., Moradkhani, H., Moghaddam Vahed, M., Poczai, P., &amp; Siddique, K. H. (2019). ipastic: An online toolkit to estimate plant abiotic stress indices.
           Applications in Plant Sciences, 7(7). https://doi.org/10.1002/aps3.11278
Sabouri, A., Dadras, A.R., Singh V.,  Azar, M., Kouchesfahani, A. S., Taslimi, M. and Jalalifar, R. (2022). Screening of rice drought‑tolerantlines by introducing a new composite selection index and competitive with multivariate methods.
           Scientific Reports, 12. https://doi.org/10.1038/s41598-022-06123-9
Fischer, R. and Maurer, R. (1978) Drought Resistance in Spring Wheat Cultivars. I. Grain Yield Responses. Australian Journal of Agricultural Research, 29, 897-912. https://doi.org/10.1071/AR9780897
}
\seealso{
\code{\link[gpbStat]{ltc}, \link[gpbStat]{ltcchk}, \link[gpbStat]{ltcmt}}
}
\author{
Nandan Patil \email{tryanother609@gmail.com}
}
