% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltcaplha.R
\name{ltcalpha}
\alias{ltcalpha}
\title{Analysis of Line x Tester data laid out in Alpha Lattice experimental design}
\usage{
ltcalpha(data, block, replication, line, tester, y)
}
\arguments{
\item{data}{dataframe containing following variables}

\item{block}{Number of blocks per replication}

\item{replication}{Number of replications}

\item{line}{Number of lines}

\item{tester}{Number of testers}

\item{y}{a dependent variable}
}
\value{
Results displayed
}
\description{
Used for analyzing the line x tester data containing only crosses laid out in alpha lattice design
}
\details{
Analyzing the line by tester data only using the data from crosses which are evaluated in alpha lattice design.
}
\seealso{
\code{aov}
}
\author{
Nandan Patil
}
