% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltcrcbd.R
\name{ltcrcbd}
\alias{ltcrcbd}
\title{Analysis of Line x Tester data laid out in RCBD experimental design}
\usage{
ltcrcbd(data, replication, line, tester, y)
}
\arguments{
\item{data}{dataframe containing following variables}

\item{replication}{Number of replications}

\item{line}{Number of lines}

\item{tester}{Number of testers}

\item{y}{a dependent variable}
}
\value{
Results displayed
}
\description{
Used for analyzing the line x tester data containing only crosses laid out in RCBD design
}
\details{
Analyzing the line by tester data only using the data from crosses which are evaluated in randomised complete block design.
}
\seealso{
\code{aov}
}
\author{
Nandan Patil
}
